refactor: make useState more universally, refactor some styles
