+++ src/components/ui/checkbox/checkbox.module.scss
@@ -2,78 +2,87 @@ button {
   all: unset;
 }
 
-.checkbox {
-  position: relative;
-
+.wrapper {
   display: flex;
+  gap: 0.5rem;
   align-items: center;
-  justify-content: center;
 
-  box-sizing: border-box;
-  width: 18px;
-  height: 18px;
+  .checkboxWrapper {
+    display: flex;
+    align-items: center;
+    justify-content: center;
 
-  background: var(--color-light-100);
-  border-radius: 3px;
+    width: 2.25rem;
+    height: 2.25rem;
 
-  &::before {
-    content: '';
-
-    position: absolute;
-    z-index: -1;
-
-    display: none;
-
-    width: 36px;
-    height: 36px;
-
+    background-color: transparent;
     border-radius: 50%;
-  }
 
-  &[data-state='unchecked'] {
-    background: transparent;
-    border: 2px solid var(--color-light-900);
-  }
-
-  .checkboxIndicator {
-    width: 100%;
-    height: 100%;
-    color: var(--color-dark-900);
-
-    &[data-state='unchecked'][data-disabled] {
-      border: 2px solid var(--color-dark-100);
+    &:hover,
+    &:hover .root:not([data-state='checked']) {
+      background: var(--color-dark-500);
     }
 
-    &[data-state='checked'][data-disabled] {
-      color: var(--color-light-100);
+    &:focus-within {
       background: var(--color-dark-100);
+    }
+
+    &:active:not(.disabled) {
+      background: var(--color-dark-100);
+    }
+
+    &.disabled {
+      background-color: transparent;
+    }
+
+    .checkbox {
+      cursor: pointer;
+
+      position: relative;
+
+      display: flex;
+      align-items: center;
+      justify-content: center;
+
+      box-sizing: border-box;
+      width: 18px;
+      height: 18px;
+
+      background: var(--color-light-100);
       border-radius: 3px;
+
+      &[data-disabled] {
+        cursor: default;
+      }
+
+      &[data-state='unchecked'] {
+        background: transparent;
+        border: 2px solid var(--color-light-900);
+      }
+
+      &[data-state='checked'][data-disabled] {
+        background: var(--color-dark-100);
+        border-radius: 3px;
+      }
+
+      .checkboxIndicator {
+        width: 100%;
+        height: 100%;
+        color: var(--color-dark-900);
+
+        &[data-state='unchecked'][data-disabled] {
+          border: 2px solid var(--color-dark-100);
+        }
+
+        &[data-state='checked'][data-disabled] {
+          color: var(--color-light-100);
+        }
+
+        .checkIcon {
+          margin-top: -2px;
+          margin-left: -2px;
+        }
+      }
     }
-
-    .checkIcon {
-      margin-top: -2px;
-      margin-left: -2px;
-    }
-  }
-
-  input[type='checkbox'] {
-    &:checked + .checkboxIndicator {
-      border: 1px solid var(--color-light-900);
-    }
-  }
-
-  &:active::before {
-    display: inline-block;
-    background: var(--color-dark-100);
-  }
-
-  &:hover::before {
-    display: inline-block;
-    background: var(--color-dark-500);
-  }
-
-  &:focus-visible::before {
-    display: inline-block;
-    background: var(--color-dark-100);
   }
 }
